Updating to Unity 5.6.0f3

+++ UnityEditor/UnityEditor/MemoryElement.cs
@@ -12,7 +12,7 @@ namespace UnityEditor
 
 		public ObjectInfo memoryInfo;
 
-		public int totalMemory;
+		public long totalMemory;
 
 		public int totalChildCount;
 
@@ -40,7 +40,7 @@ namespace UnityEditor
 			this.expanded = false;
 			this.memoryInfo = memInfo;
 			this.name = this.memoryInfo.name;
-			this.totalMemory = ((memInfo == null) ? 0 : memInfo.memorySize);
+			this.totalMemory = ((memInfo == null) ? 0L : memInfo.memorySize);
 			this.totalChildCount = 1;
 			if (finalize)
 			{
@@ -53,7 +53,7 @@ namespace UnityEditor
 			this.name = n;
 			this.expanded = false;
 			this.description = "";
-			this.totalMemory = 0;
+			this.totalMemory = 0L;
 			this.totalChildCount = 0;
 			this.children = new List<MemoryElement>();
 			foreach (MemoryElement current in groups)
